Compare attribute codes using lowercase
Attribute option codes are prefixed with the attribute code to prevent clashes. This bit of code will attempt to remove it before saving it to magento.
We've recently introduced support for uppercase attributes, which are just lowercased (See: https://github.com/snowio/akeneo3-data-model/pull/4) so we add that check here as well.